# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_thienel_go-backend-template_internal_interface_api_dto.CMDPlan:
    properties:
      device_configured:
        type: boolean
      manifest_urls:
        items:
          type: string
        type: array
      profile_names:
        items:
          type: string
        type: array
    type: object
...
  github_com_thienel_go-backend-template_internal_interface_api_dto.EventSubscription:
    properties:
      context:
        type: string
      event:
        type: string
      event_context:
        type: string
      workflow:
        type: string
    type: object
...
  github_com_thienel_go-backend-template_internal_interface_api_dto.NanoCMDInventoryDevice:
    properties:
      model:
        type: string
      serial_number:
        type: string
    type: object
  github_com_thienel_go-backend-template_internal_interface_api_dto.NanoCMDInventoryResponse:
    additionalProperties:
      $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.NanoCMDInventoryDevice'
    type: object
  github_com_thienel_go-backend-template_internal_interface_api_dto.NanoCMDProfile:
    properties:
      identifier:
        type: string
      uuid:
        type: string
    type: object
...
  github_com_thienel_go-backend-template_internal_interface_api_dto.NanoCMDWebhook:
    properties:
      acknowledge_event:
        additionalProperties: {}
        type: object
      checkin_event:
        additionalProperties: {}
        type: object
      created_at:
        type: string
      event_id:
        type: string
      topic:
        type: string
    type: object
  github_com_thienel_go-backend-template_internal_interface_api_dto.NanoCMDWorkflowStartResponse:
    properties:
      instance_id:
        type: string
    type: object
...
  github_com_thienel_go-backend-template_pkg_response.APIResponse-any:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.Error'
      is_success:
        type: boolean
      message:
        type: string
    type: object
...
  ? github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_CMDPlan
  : properties:
      data:
        $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.CMDPlan'
      error:
        $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.Error'
      is_success:
        type: boolean
      message:
        type: string
    type: object
...
  ? github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_EventSubscription
  : properties:
      data:
        $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.EventSubscription'
      error:
        $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.Error'
      is_success:
        type: boolean
      message:
        type: string
    type: object
...
  ? github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_NanoCMDInventoryResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.NanoCMDInventoryResponse'
      error:
        $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.Error'
      is_success:
        type: boolean
      message:
        type: string
    type: object
...
  ? github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_NanoCMDWorkflowStartResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.NanoCMDWorkflowStartResponse'
      error:
        $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.Error'
      is_success:
        type: boolean
      message:
        type: string
    type: object
...
  ? github_com_thienel_go-backend-template_pkg_response.APIResponse-map_string_github_com_thienel_go-backend-template_internal_interface_api_dto_NanoCMDProfile
  : properties:
      data:
        $ref: '#/definitions/map_string_github_com_thienel_go-backend-template_internal_interface_api_dto.NanoCMDProfile'
      error:
        $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.Error'
      is_success:
        type: boolean
      message:
        type: string
    type: object
  github_com_thienel_go-backend-template_pkg_response.Error:
    properties:
      code:
        type: string
      fields:
        items:
          $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.FieldError'
        type: array
      message:
        type: string
    type: object
  github_com_thienel_go-backend-template_pkg_response.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
...
  map_string_github_com_thienel_go-backend-template_internal_interface_api_dto.NanoCMDProfile:
    additionalProperties:
      $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.NanoCMDProfile'
    type: object
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Server for THD iOS MDM Portal.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: THD iOS MDM Portal API
  version: "1.0"
paths:
...
  /api/v1/nanocmd/cmdplan/{name}:
    get:
      description: Retrieve and return a named command plan as JSON.
      parameters:
      - description: User-defined name of Command Plan.
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_CMDPlan'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Retrieve and return a named command plan
      tags:
      - NanoCMD Setup
    put:
      consumes:
      - application/json
      description: Upload a named JSON command plan.
      parameters:
      - description: User-defined name of Command Plan.
        in: path
        name: name
        required: true
        type: string
      - description: Command plan.
        in: body
        name: plan
        required: true
        schema:
          $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.CMDPlan'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Upload a named command plan
      tags:
      - NanoCMD Setup
  /api/v1/nanocmd/event/{name}:
    get:
      description: Retrieve the event subscription.
      parameters:
      - description: User-defined name of Event Subscription.
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_EventSubscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Retrieve the event subscription
      tags:
      - NanoCMD Setup
    put:
      consumes:
      - application/json
      description: Store the event subscription provided in the request body.
      parameters:
      - description: User-defined name of Event Subscription.
        in: path
        name: name
        required: true
        type: string
      - description: Event Subscription.
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.EventSubscription'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Store the event subscription
      tags:
      - NanoCMD Setup
  /api/v1/nanocmd/fvenable/profiletemplate:
    get:
      description: Returns the FileVault enable Configuration Profile template.
      produces:
      - application/x-apple-aspen-config
      responses:
        "200":
          description: Apple Configuration Profile
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Returns the FileVault enable Configuration Profile template
      tags:
      - NanoCMD Setup
  /api/v1/nanocmd/inventory:
    get:
      description: Retrieve inventory data for enrollment IDs.
      parameters:
      - collectionFormat: csv
        description: Enrollment ID. Unique identifier of MDM enrollment. Often a device
          UDID or a user channel UUID.
        in: query
        items:
          type: string
        name: id
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_NanoCMDInventoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Retrieve inventory data for enrollment IDs
      tags:
      - NanoCMD Setup
  /api/v1/nanocmd/profile/{name}:
    delete:
      description: Deletes the named profile.
      parameters:
      - description: User-defined name of Profile.
        in: path
        name: name
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Deletes the named profile
      tags:
      - NanoCMD Setup
    get:
      description: Fetches the named raw profile.
      parameters:
      - description: User-defined name of Profile.
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/x-apple-aspen-config
      responses:
        "200":
          description: Apple Configuration Profile
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Fetches the named raw profile
      tags:
      - NanoCMD Setup
    put:
      consumes:
      - application/x-apple-aspen-config
      description: Uploads a raw profile. Signed profiles also supported.
      parameters:
      - description: User-defined name of Profile.
        in: path
        name: name
        required: true
        type: string
      - description: Raw profile mobileconfig.
        in: body
        name: data
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Uploads a raw profile
      tags:
      - NanoCMD Setup
  /api/v1/nanocmd/profiles:
    get:
      description: Retrieve profile metadata.
      parameters:
      - collectionFormat: csv
        description: User-defined name of profile.
        in: query
        items:
          type: string
        name: name
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-map_string_github_com_thienel_go-backend-template_internal_interface_api_dto_NanoCMDProfile'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Retrieve profile metadata
      tags:
      - NanoCMD Setup
...
  /api/v1/nanocmd/webhook:
    post:
      consumes:
      - application/json
      - application/json
      description: |-
        Handler for MicroMDM-compatible webhook callback.
        Endpoint to receive MDM/CMD check-in events and state updates
      parameters:
      - description: Webhook payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.NanoCMDWebhook'
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      summary: NanoCMD Webhook
      tags:
      - NanoCMD Setup
      - Infrastructure
  /api/v1/nanocmd/workflow/{name}/start:
    post:
      description: Start a workflow.
      parameters:
      - description: Name of NanoCMD workflow.
        in: path
        name: name
        required: true
        type: string
      - collectionFormat: csv
        description: Enrollment ID. Unique identifier of MDM enrollment. Often a device
          UDID or a user channel UUID.
        in: query
        items:
          type: string
        name: id
        required: true
        type: array
      - description: Workflow-dependent context.
        in: query
        name: context
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_NanoCMDWorkflowStartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Start a workflow
      tags:
      - NanoCMD Setup
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...